feat(nftables): make sets accept CIDRs, remove dynamic flag
The interval flag allows adding IP ranges in CIDR notation, e.g.:

nft add element inet sesame allowed_clients { 10.0.2.0/24 . 50000 timeout 1h }

This is not useful now, but I want to leave the door open to allowing
access to blocks of IPs (e.g. entire countries), so the CIDR notation is
more flexible, and I can always target specific IPs with /32 or /128.

The dynamic flag conflicts with interval, for some reason, and
specifying both fails with a cryptic "Could not process rule: Operation
not supported" error, so I had to remove it.
The dynamic flag is only needed if the set will be modified during
packet processing, e.g. for rate limiting, etc. But for my use case the
set will be modified manually, so I can live without it.