feat: make default access duration configurable
Previously this was hardcoded in the NFTables implementation, and the
Manager had no knowledge of it, which caused showing an incorrect
duration in log messages when `open --access-duration` wasn't specified
(0), even though it was actually set to the default NFTables timeout value (5m).

This way the duration is passed to the firewall via the manager, and
it's also a configurable value, either when `init` is run, or in the
configuration file. Though changing it in the configuration file after
`init` was run is currently not tested. This needs proper testing, of
course.