feat(cli): add user command
This was mostly copied from Disco. The only differences are the removal
of the `user update` subcommand, since there is currently no metadata
associated to a user that needs to be updated, and the upgrade to the
tablewriter v1 library, which has an entirely different API (for the
better :).

Users are currently very barebones, since I'm not planning to support
roles, like Disco has. But I still want to keep them, since invites need
to be associated with users, and in the future it will likely be
necessary to track actions performed by users, as a sort of audit log.

Access control still might eventually be useful, but I'll leave that
complexity for when it's actually needed.

The tablewriter dependency seems overkill for the basic rendering I'm
doing now, but it will surely come in handy later. I want to eventually
output firewall rules, for example, and the rendering for that will be
more complex.