feat(app): unwrap joined RuntimeError
It seems that at some point the error returned by app.Run is joined
using errors.Join, but I can't trace where it happens. So even though
all commands return the internal RuntimeError, Errorf wasn't actually
extracting its fields.

I initially wrote this for Disco so that the logged messages would be
friendlier, and easier to parse.

So that instead of:

2025-07-06 15:07:00.700 ERR failed removing user: user with name 'test' doesn't exist

Users would see:

2025-07-06 15:10:02.344 ERR failed removing user cause="user with name 'test' doesn't exist"

I suppose the colon arguably implies that the second message is the
cause for the first one, but using a "cause" label makes that explicit.
And since we're already using slog, it's more idiomatic to use distinct
fields.
The RuntimeError is also useful when it includes a "hint" field, which
I'm not currently using, but it will come in handy later.

I'll probably eventually want to make this use the causer interface[1]
added in recent versions of Go, which didn't exist in 2024 when I
originally wrote this, but this works for now.

[1]: https://pkg.go.dev/github.com/pkg/errors#readme-retrieving-the-cause-of-an-error