feat(db): pass expiration to NewInvite, set created/updated in Save
I will change the CLI to receive an expiration time instead of duration,
since it's more flexible.

I also prefer to create the created/updated timestamps in Save, instead
of externally. This makes it less likely for the caller to forget to set
them. And created/updated is really the time when the record was stored
in the DB, not when the instance was created in memory, so this way is
slightly more accurate.