refactor(tls): more robust and simpler NewTLSCert
I hope this is the final form of this function...

It now only returns a tls.Certificate, which contains both the x509 cert
and the private key.

It also does some additional error checking, and uses the new helper
function added in the previous commit for extracting the CA cert.

This was done with the help of Claude 4 Sonnet, and I understand the Go
TLS APIs a bit better now. :)