feat(web): add remote join API endpoint handler
This is a slightly different and more robust version of the protocol
used in Disco.

The main difference is that the client authentication is not done by
verifying the signature from ephemeral Ed25519 keys, but by verifying an
HMAC using the shared ECDH key. This is simpler, and HMAC is more
appropriate for this use case.

The other differences are related to how the TLS certs are stored and
returned, the payload itself, and the use of some helper TLS functions.

This was done with the assistance from Claude 4 Sonnet. I do think the
current version is an improvement, even though the one in Disco is not
technically insecure.