feat(web): add hybrid TCP/TLS listener
This is somewhat hacky, but it allows the server to serve both
unencrypted HTTP and HTTPS from a single port. This simplifies the setup
on server, reducing the attack surface, and the usage on the client.
I.e. instead of requiring one address for authenticating/joining (adding
a remote), and another for mTLS authenticated usage, a single address
can be reused.