feat(web): use Chi middleware chaining API
This removes the bespoke chaining API I was using before. The Chi
variant is not as ergonomic, as it doesn't allow passing both handlers
and middleware, but their approach doesn't rely on runtime type
checking, and is arguably more intuitive.

And since I'm already depending on their middleware code, this doesn't
add an additional dependency.