feat: add redeemed_at field to Invite, mark as redeemed
This ensures that invite tokens are not accepted more than once.

My poor man's query builder is a bit wonky since I don't have a good
abstraction for handling null values in queries, but it happens to work
with time.Time since the zero value is unusable.

I did briefly consider other options, including adopting a more robust
and established query builder, but ultimately decided to leave this as
is, and focus on more important things. There are many lightweight query
builders for Go, and Bob[1] seems like the most promising. Migrating to
something else at this point would be a substantial amount of work for
no immediate gain.

[1]: https://bob.stephenafamo.com/