feat: use time from app context instead of system time
These are some leftover places where the system time was still being
accessed directly, instead of using the app context time which can be
mocked.

The only two places where the system time is referenced is in main(),
for obvious reasons, and in the AppRemoteIntegration test because TLS
verification requires it.