feat(cli,db): add remote removal
The remote integration test is getting to be a bit too bloated for my
taste. I would prefer to split it into smaller tests that could still be
integration-y, without testing as much.

The problem with remotes specifically is that the only way to add them
is by having a valid remote token. Otherwise a lot of complex mocking
would need to happen, or I would need to introduce another app option
to specifically mock things. Which I don't want to do. I don't like
adding code that only serves tests...

So I'll leave this for now, but will think about some alternatives.