refactor(firewall): make Firewall implementation receive IPSet
Previously, individual IP ranges were added one by one to the
nftables set, but the Go API allows adding multiple elements to the set
at once. This reduces the Netlink load when granting access to multiple IP
ranges. So passing an IPSet instead of an IPRange is a straightforward
change.

I'm still not sure what the value of the Manager is, and I considered
removed it altogether. I suppose it's a layer between Sesame packages
and Firewall implementations, since Firewalls don't have to know about
Services, and other Sesame internals. This is why the API is slightly
different (GrantAccess vs Allow), which also simplifies it from the
perspective of other Sesame packages. In the future it also might be
convenient to add more Sesame logic to it, such as how services are
handled, etc. So I decided to leave it after all.