feat(db): add Invite.SiteID field
This is a server-side reference for the remote site. Currently it's only
associated to the invite, but I plan to add tracking of issued TLS
client certificates next, so this will be used as a user-friendly
identifier of the remote client, and indirectly of the specific client
certificate chain.

It will also be useful in audit logs so that we can distinguish the
particular site/remote the action originated from, in addition to the
user.