feat(cli): list invites by status, allow filtering
This makes the invite list output much more useful and flexible.

A Status column was added that makes it clear what the status is of each
invite, sorted by the most helpful criteria in each case. It's also
possible to filter by status, and the previous `--all` flag now becomes
a special value of `--status`.

Additionally, tokens are now truncated by default, mostly because
they're too long and mess up the table output, but also for security
reasons. A `--token` flag was added to output them in full as before.