!feat(web): add composable handler package
This required quite a few iterations to get right, with copious cursing
at Go's limited type system and half-baked generics implementation...

(╯°□°)╯︵ ┻━┻

Like, seriously, I get that there might be deep technical reasons why
type parameters can't be declared on methods (which I won't pretend to
understand), but after 4 years of heated discussion[1] one would think
that it would be clear to maintainers that this is a highly requested
feature, and that there could be *some* possible implementation that
doesn't require road-blocking trade-offs. The reactions I've seen range
from "we're not against it in principle, it's just that nobody has
proposed an implementation that's up to our standards" to "this is
impossible with how Go currently works".

Ugh.

There are other head-scratchers, like being unable to define union types
with behavioral interfaces[2], and the overall unreadable syntax that
ruined what was once a simple and legible language.

So, despite all my failed efforts, this is the best I could do to arrive
at a user-friendly and composable API for building HTTP handlers. It
does require the use of reflection, and hidden dependencies passed via
the request context, but ultimately, I'm quite happy with it.

The results speak for themselves: the previous `/join` handler was 120
lines, most of which was authentication and request/response processing.
The current version is 14 lines dedicated to the core business logic:
creating a client TLS cert. The request and response are properly typed,
and everything not unique to this endpoint is offloaded to the handler
helper package.

The helper package, in turn, is generic and composable, which means that it
can be reused by other endpoints, and composed or extended in different
ways. This makes it trivial, for example, to change specific endpoints to
speak protobuf instead of JSON, or to implement custom functionality.
Considering that the non-TLS Sesame endpoints use a bespoke protocol for
authentication and encryption, this is very valuable. I plan to add a
`/renew` endpoint soon which will use a similar protocol as `/join`,
just with a different token auth, which was a big motivator for the
reusability aspect. The fact that the package handles errors
transparently is also a big win.

All in 290 SLOC! \o/

This does need more thorough testing, of course, but it currently passes
my broad remote integration test, and it worked as expected with manual
testing, so I'm happy with it for now.

It's currently somewhat related to Sesame internals, but I don't see why
this couldn't be a standalone package that can be used by any Go server app.
I know that I'm copying it into all of my Go web projects moving
forward.

Full disclosure: this was done with assistance from Claude Sonnet 4,
though to be honest, Claude was a hindrance most of the time. It
routinely got confused by Go's type system and produced code that
wouldn't compile. It was helpful during design phases as a slightly
smarter rubber duckie, and with some of the chore tasks like writing
documentation, but other than that, it certainly wasn't the 10x
productivity booster AI companies love to promote. Most of the features
were designed and implemented by human hands. There is no way that it
could produce code of this quality on its own, no matter how good the
prompting is. So artisan programming is alive and well in the age of
"AI" (◡ ‿ ◡ .)

[1]: https://github.com/golang/go/issues/49085

[2]: https://github.com/golang/go/issues/49054