feat(web): send client pub key in Authorization header for join endpoint
This changes the authentication scheme for the join endpoint to send the
client's X25519 pubkey in the Authorization header instead of the the
body.

This is done in order to be able to standardize the implementation of
API handlers on the server side. Previously the join endpoint was
special in the sense that the request was expected to contain raw data
(the client's pubkey), while the response would be returned as JSON.

This discrepancy between request and response made it difficult to
standardize on a "serializer" abstraction, which in turn made it
difficult to properly implement request validation.

Since both the invitation token and client's pubkey are required for
proper join authentication, it makes sense for both to be sent in the
Authorization header, and for both the request and response bodies to
contain JSON data. Even though the join request data itself is empty,
the standardization itself makes things easier on the server side.