feat(web): add Serializer abstraction to handler package
Previously the JSON marshaller and unmarshaller functions were "special"
request and response processors in that they were the only ones that
converted raw data into a typed value. Treating them as generic request
and response processors made it impossible to properly implement request
validation, since validation should happen after serialization but
before request processing (i.e. validation should be done on the typed
request value).

Introducing a Serializer abstraction makes it possible to do request
validation before the deserialization is run.

This does limit the utility of the handler package to pipelines with
endpoints that expect the same input and output format, but this is
generally acceptable for API endpoints. And deviating from this standard
could be done by implementing a Serializer variant with no-op or mixed
Serialize and Deserialize methods.