feat(web/server): add request validation
This adds support to the handler package for requests to have a
Validate() method, and implements request validation for all current
endpoints.

This specific feature was simple to add, but it required changing the
authentication scheme of the join endpoint, and introducing a Serializer
abstraction, as done in previous commits. This is because request
validation should happen *after* successful authentication AND
deserialization, but *before* request processing.

I'm quite happy with how it turned out in the end. :)