feat(web/client): extract error message from API responses
This was tricky to do, especially for the /join endpoint, but from my
manual tests it works as expected.

Essentially, if a request fails for any reason, I want to show a useful
error message to the CLI user. The /join endpoint complicates this a bit
since the response could be encrypted or unencrypted, so handling both
scenarios was a bit difficult.